3.2-crud/stocks_products/requirements.txt
django
djangorestframework
psycopg2-binary
django_filter
python-dotenv
gunicorn
flake8